+++ js/app.js
@@ -9,12 +9,6 @@ Behaviours.add( 'media:sliders', () => {
     Behaviours.add( 'media:sliders', doSliders );
 } );
 
-// This event is triggered once a slider is initialized.
-// Do any slider adjustment here!
-window.addEventListener( 'swiper:afterInit', e => {
-    const swiper = e.detail;
-}, { passive: true } );
-
 // Close CCM widget on privacy policy page.
 Behaviours.add( 'gdpr:privacy-policy', $context => {
 
@@ -23,4 +17,51 @@ Behaviours.add( 'gdpr:privacy-policy', $context => {
 
     CCM.closeWidget()
 
-} )
+} );
+
+// This event is triggered once a slider is initialized.
+// Do any slider adjustment here!
+window.addEventListener( 'swiper:afterInit', e => {
+    const swiper = e.detail;
+}, { passive: true } );
+
+/**
+ * Toggle header.
+ */
+(function() {
+
+    const threshold = 11 // min distance to be scrolled before making actions
+    let $masthead;
+
+    Behaviours.add( 'header:toggle', $context => {
+
+        let $primary = $context.querySelector( '#primary' )
+        if ( !$primary && $context?.id === 'primary' )
+            $primary = $context;
+
+        if ( $primary ) {
+            if ( ($masthead && $masthead.parentElement ) || ($masthead = document.getElementById( 'masthead' )) )
+                $masthead.addEventListener( 'mouseover', e => $masthead.classList.add( 'is-visible' ) )
+        }
+        // else $masthead = undefined
+
+    } )
+
+    let lastScrollY = 0;
+    window.addEventListener('scroll', e => {
+        if ( Math.abs( window.scrollY - lastScrollY ) < threshold ) return
+
+        if ( $masthead ) {
+            $masthead.classList[window.scrollY < $masthead.offsetHeight ? 'remove' : 'add']( 'is-togglable' );
+
+            const action = lastScrollY < window.scrollY
+                ? /* scroll up */ 'remove'
+                : lastScrollY > window.scrollY ? /* scroll down */ 'add' : ''
+
+            if ( action ) $masthead.classList[action]( 'is-visible' );
+        }
+
+        lastScrollY = window.scrollY
+    }, false);
+
+})();