+++ utilities/inc/delete-revisions.php
@@ -1,0 +1,8 @@
+<?php
+
+//foreach ( get_posts( array(
+//	'post_type' => 'revision',
+//	'post_status' => 'inherit',
+//	'posts_per_page' => -1,
+//	'fields' => 'ids',
+//) ) as $revision_id ) wp_delete_post_revision( $revision_id );