+++ utilities/inc/pages-for-.php
@@ -106,7 +106,7 @@ function page_for_class( $classes ) {
 }
 
 // get bogo translated post
-add_action( 'wp', function() {
+if ( function_exists( 'bogo_get_local_post' ) ) add_action( 'wp', function() {
 	foreach( apply_filters( 'pages_for', array() ) as $name => $label ) {
 		add_filter( "option_$name", 'bogo_get_local_post' );
 	}